Rename backstore/store.rs to backstore/backstore.rs
For more consistent naming.
