+++ src/engine/strat_engine/backstore/mod.rs
@@ -2,6 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+#[allow(module_inception)]
+mod backstore;
 mod blockdev;
 mod blockdevmgr;
 mod cleanup;
@@ -9,9 +11,8 @@ mod device;
 mod metadata;
 mod range_alloc;
 mod setup;
-mod store;
 mod util;
 
 pub use self::metadata::MIN_MDA_SECTORS;
 pub use self::setup::{find_all, get_metadata, is_stratis_device};
-pub use self::store::Backstore;
+pub use self::backstore::Backstore;
